test: make smoke_test less flaky
The `--int` test was flaky because it wasn't waiting for the job to
actually start before trying to interrupt, and so sometimes it would
fail the single `--int` call because the job had not yet started.

+++ jobrunner/info.py
@@ -330,6 +330,9 @@ class JobInfo(object):
         parent.inactive[k] = self
 
     def _killPgrp(self):
+        if not self.pid:
+            print('not running')
+            return
         pgrp = os.getpgid(self.pid)
         sig = signal.SIGINT
         os.killpg(pgrp, sig)
